requirements.txt
certifi==2023.7.22
contourpy @ file:///D:/bld/contourpy_1686734008624/work
CoolProp==6.4.1
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
fonttools @ file:///D:/bld/fonttools_1689965632368/work
future==0.18.3
importlib-resources @ file:///home/conda/feedstock_root/build_artifacts/importlib_resources_1689017639396/work
kiwisolver @ file:///D:/bld/kiwisolver_1666805897768/work
matplotlib @ file:///D:/bld/matplotlib-suite_1678135799522/work
munkres==1.1.4
numpy @ file:///D:/bld/numpy_1688887186903/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
Pillow @ file:///D:/bld/pillow_1688255909150/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1690737849915/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
rocketcea==1.1.34
scipy==1.11.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
unicodedata2 @ file:///D:/bld/unicodedata2_1667240049903/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1689374466814/work
